Make tests pass on hosts that have no ipv4 connectivity
Hosts with no ipv4 connectivity are going to be increasingly common in
the future. In such hosts, one will see errors like the following when
running the rack tests:

Errno::EADDRNOTAVAIL: Cannot assign requested address - bind(2) for 127.0.0.1:9210

Binding to localhost instead of explicitly to
127.0.0.1 solves the problem - `localhost` is resolved to all
_available_ local addresses and those are used.  Client connections
need to do the same, and it all just works regardless of the local
networking stack.