Encrypted session cookies using Rack::Encryptor
Add Rack::Encryptor which uses AES-256-CTR with HMAC-SHA-256 for
authentication. The secret key must be 64 bytes long and the first 32
bytes are used for the base encryption key and the remaining bytes are
used for the authentication code. A random value is generated at
encryption time to create a per-message encryption key. This key is
generated using HMAC-SHA-256 and the initial base secret key.

For a complete history of these changes please see:
- https://github.com/rack/rack/pull/1177
- https://github.com/mjc-gh/rack/tree/rack-aead-session-cookie-pr-1177-archive