Fix MockRequest cookie parsing for hash with array value
Problem:
`Rack::Response.new` allows to set multiple cookies just by passing hash with key 'Set-Cookie' and value as an array containing cookies we want to set.
example:
```ruby
Rack::Response.new(
  [ "Authentication failed" ],
  401,
  {
    "Content-type" => "text/error",
    "Set-Cookie" => ["foo=bar", "baz=ban"]
  }
)
```
Unfortunately `MockRequest` class is not able to parse such definition of cookies as it expects cookie value to be always a string.

Solution:
Wrap every cookie value to `Array()` which converts string values to array. So we can safely iterate over the values. In case the value is already an array it simply return it.