Parse 'expires' to Time in MockResponse cookies
CGI::Cookie expects the 'expires' attribute, if set, to be an instance
of Time.  While this was being correctly set for 'max-age', 'expires'
was set as-is, a String.  This caused errors when inspecting the
response, as CGI::Cookie#inspect calls #gmtime on the 'expires' value.

This commit fixes MockResponse to parse 'expires' as an RFC1123
formatted date/time when parsing cookies from the header.

Fixes #1758.