Make QueryParser#normalize_params API more compatible with Rack 2
Switching the depth argument to be the starting depth is not compatible
and results in breakage in rack-test (and likely other libraries).

Fix this by keeping the normalize_params API the same, and moving the
implementation to a private _normalize_params method.  Remove
now unnecessary depth arguments when calling normalize_params, and
call _normalize_params directly for internal calls.