Fix "reject insanely long boundaries" test hang
At least on OpenBSD, this test occasionally hangs, because the
`wr.write` does not return/raise after the `rd.close` in the other
thread.  Switch to `write_nonblock` with `exception: false`, using
Thread.pass if the write would block.

With this change, the test takes less than two seconds and does
not hang.