Remove unnecessary conditional in rack/static.rb
Mime.mime_type always returns a value.  The only way it wouldn't
is if MIME_TYPES was modified to have an entry with a nil/false
value, which would seem to violate the expectations of
Mime.mime_type, which is documented to return a string if the
mime type is found.