Update docs and remove a dead link.
This PR does two thinks. First, it replaces references to the
master branch with main (which seems to be the default now).

It also removes a bad link in a comment as a reference for
not parsing CGI ISINDEX parameters. I could not find a
viable alternative link to defend this assertion. If there is
one I'd be happy to add it.