Fix Utils.build_nested_query to URL-encode all query string fields
In the formatted query string, the field should be fully URL-encoded.
The previous implementation left the square brackets as un-escaped.

It was noticed that if params were provided as a nested hash it produced
a different result than a hash of string fields. For example, the params

    { 'q[s]' => 'name' }

would result in the query string:

    q%5Bs%5D=name

but

    { q: { s: 'name' } }

would result in:

    q[s]=name

Both forms now produce the same correct result with fully URL-encoded
params.