Make QueryParser::Params a Hash subclass
Because of the to_params_hash method, this cannot be Hash itself.
We should consider deprecating that method and maybe the Params
constant as well.

Most of the spec changes are just changing overrides of
Params#initialize.

This tweaks the "not create infinite loops with cycle structures"
spec, because it wasn't actually testing for loops, it was
checking the string representation was the same between Hash
and Params.  This spec could probably be eliminated, but the
tweak allows it to pass by checking that the level 1 hash string
representation is the same as the level 2 hash string representation.