Make query parameters without = have nil values (#2059)
* Revert "Prefer to use `query_parser` itself as the cache key. (#2058)"

This reverts commit 5f90c33e4ccee827cb5df3d8854dc72791345c51.

* Revert "Fix handling of cached values in `Rack::Request`. (#2054)"

This reverts commit d25feddcbe634d95ec693bfbd710167a11c74069.

* Revert "Add `QueryParser#missing_value` for handling missing values + tests. (#2052)"

This reverts commit 59d9ba903fdb50cf8db708c8263a7b2a79de83fb.

* Revert "Split form/query parsing into two steps (#2038)"

This reverts commit 9f059d19647aeaef5c2cc683a333c06120caf939.

* Make query parameters without = have nil values

This was Rack's historical behavior.  While it doesn't match
URL spec section 5.1.3.3, keeping the historical behavior avoids
all of the complexity required to support the URL spec standard
by default, but also support frameworks that want to be backwards
compatible.

This keeps as much of the specs added by the recently reverted
commits that make sense.
# Conflicts:
#	lib/rack/multipart.rb
#	lib/rack/request.rb
#	test/spec_request.rb