Ignore Range header if served file is 0 bytes (#2159) (#2159)
Normally Rack::Files truncates a byte range to fit a file's contents as long as
at least some bytes are within range. However, an empty (0 byte) file is a
special case. Previously, empty files requested with a Range header always
resulted in a 416 error. This changes 0 byte files to ignore the Range header
and return the empty file.

This improves compatibility with clients that speculatively request files with
byte ranges.