Remove deprecated base64 in favour of pack/unpack
The base64 library is external as of Ruby 3.4 and is no longer available
by default in Ruby. The pack/unpack methods support encoding as base64
and can be used in it's place.

String#unpack.first was preferred here to support Ruby 2.3, which didn't
yet have String#unpack1 introduced in Ruby 2.4.