Deprecate automatic cache invalidation in Request#{GET,POST}
This warns in Request#GET if the current query string value does not
match the query string value used when setting
RACK_REQUEST_QUERY_HASH (similar for Request#POST
and the request body).  In Rack 3.2, there will be no warning,
it will use the cached value instead of reparsing in these cases.

It appears as though env[RACK_REQUEST_FORM_VARS] is already
unnecessary, as the value is set but never accessed, dating back to
its introduction in 6c80c6cf86f1f897e08882cb6c9ae731fab1b9c1.
However, even though it is never used by Rack, it apparently is
used by Rails.  However, Rails usage appears to be limited to
parameter filtering, and if the RACK_REQUEST_FORM_VARS key wasn't
set, there would be nothing to filter.  So it's possible Rails
could be changed so that if the key was missing, there are no
problems (maybe it works like that already, and only the Rails
tests need updates).