Fix encoding setting for non-binary IO-like objects in MockRequest#env_for
No core class implements both encoding and set_encoding.

Classes implementing encoding:

* Regexp
* Symbol
* String

Classes implementing set_encoding:

* File
* ARGF.class
* IO

This code was added in 64610db2d2e7e910c4d3e17874dc316ed3eb227a.
Previously, set_encoding was always called.
64610db2d2e7e910c4d3e17874dc316ed3eb227a actually broke encoding
setting to binary for IO objects, but because there were no tests
for the case, it wasn't caught.

I don't see a point of warning of Regexp and Symbol use, and
raising for objects that support encoding but not set_encoding
doesn't make sense.  I'm guessing the intention was to handle
IO-like objects, and then internal_encoding could be checked,
but it's simpler to just call set_encoding in that case, as
that was the previous behavior.

An alternative to this code would be reverting
64610db2d2e7e910c4d3e17874dc316ed3eb227a, but that is no longer
backwards compatible with 3.1.0, which allows objects that do
not respond to set_encoding.

Found by coverage testing.