Do not remove escaped opening/closing quotes for content-disposition filenames
This regexp was already questionable because of the use of ^ and $.
It turns out this was no longer needed, since the parser earlier
in the same method handles the quotes, so the result is that this
was removing escaped quotes that were part of the filename.

Found by coverage testing.