Resolve deprecation warnings about uri `DEFAULT_PARSER` (#2242)
Ruby switches the default parser from RFC2396 to RFC3986. This parser contains
all methods from the old parser but delegates to RFC2396 for a few and warns.
Namely `extract`, `make_regexp`, `escape`,and `unescape`.

Since the RFC2396 alias is not available for old rubies, do a defined check and select the appropriate one.
This will work for all versions without warnings.