fix: malformed charset param (#2263)
If a content type header is provided with a malformed part, there's no
nice way of interpreting it since instead of being a key/value pair it's
just a key.

Instead, we mimic the behaviour of query parameters and opt to return an
empty string in lieu of missing values—this way we can still identify
their presence, but won't blow up. The behaviour isn't identical, but
we're also not worried about backwards compatibility here (given the
alternative was exploding)