Freeze various constants
This ensures these constants can't be mutated at runtime and also allows
them to be used inside Ractors

I tried to be very conservative in limiting this to constants users
should not be modifying. For example FORM_DATA_MEDIA_TYPES,
PARSEABLE_DATA_MEDIA_TYPES, MIME_TYPES are documented to be modifyable
by users.