Fix streaming bodies when using `Rack::Events`
When `Rack::Events` is used, it would use `EventedBodyProxy` to wrap
the body. Unlike `BodyProxy`, `EventedBodyProxy` always implements
`#each`.

As per the spec, streaming bodies must both implement `#call` and not
implement `#each`, and if both are present, the server must call
`#each`, not `#call`.

When a streaming body was wrapped by an `EventedBodyProxy`, the proxy
will nonetheless `.respond_to?(:each)`. This would cause the server to
call `#each` on the proxy. The proxy would then forward the call to
`#each` to the body that it proxies, which does not implement `#each`,
raising a `NoMethodError`.

To fix this, `EventedBodyProxy` will only `.respond_to?(:each)` if
the body it proxies responds to `#each`.

Note that, while this fixes the `NoMethodError`, this means that the
`on_send` handler will never be invoked for streaming bodies.