Allow dynamic headers for Rack::Files
This PR adds support for dynamic headers in Rack::Files via the headers
configuration.

For example the following configuration add the header "x-dynamic" with
a random integer between 1 and 10 for each request handled by Rack::Files:

```
Rack::Files.new(".", {
  "x-dynamic" => ->(request) { rand(1..10).to_s }
})
```

This feature could be used to provide different cache durations
for different files. E.g. a longer duration for files with a
contenthash and shorter duration for files without a contenthash. This
would allow solving issues like
https://github.com/rails/rails/pull/55249 within Rack::Files directly.