Fix denial of service vulnerbilties in multipart parsing
Two separate vulnerabilities:

1. Unbounded buffering of uploaded data waiting for a boundary.

2. Unbounded buffering of uploaded data waiting for complete
   mime part header.

The respective limits are 16KB for (1) and 64KB for (2), but those
limits only apply for non-default buffer sizes. If left at the
default configuration, 1MB (default buffer size) will be the limit
for both.

This changes one EmptyContentError exception to an Error exception,
but EmptyContentError is probably the wrong error to raise for a
very long boundary.