Document that Rack::Files and Rack::Directory follow symlinks
This is the default behavior for most webservers, so it should
not be surprising. It's also not a security issue.  As
explained by the Apache documentation:

```
Omitting this option [FollowSymLinks] should not be considered
a security restriction, since symlink testing is subject to
race conditions that make it circumventable.
```

However, we have had multiple security vulnerability reports
about this behavior. To hopefully avoid future similar reports,
it is useful to explicitly document that following symlinks is
expected behavior.